+++ arduino/src/StepperMotor/StepperMotor.hpp
@@ -22,7 +22,7 @@ public:
     void start();
     void stop();
 
-    uint64_t getNumSteps() const;
+    int getNumSteps() const;
 
 private:
     pins::pin_t dir_pin_;
