+++ arduino/src/cmds/platform.cpp
@@ -112,7 +112,7 @@ String set_platform_overwrite(const String& mode)
 
 String read_platform_steps(const String& /*unused*/)
 {
-    return String((unsigned long)lowering_platform_left.getNumSteps()) + " + " + String((unsigned long)lowering_platform_right.getNumSteps());
+    return String(lowering_platform_left.getNumSteps()) + " + " + String(lowering_platform_right.getNumSteps());
 }
 
 void init_platform()
